poi_ingest/requirements.txt
# Production dependencies
django>=5.0,<6.0
djangorestframework>=3.14.0
pydantic>=2.0
python-dotenv>=1.0.0

# Development dependencies  
django-debug-toolbar>=4.2
pytest>=7.4
pytest-django>=4.5
factory-boy>=3.3
coverage>=7.3